extra doc checks
